general adjustments

+++ src/group/application/services/leave-group.service.ts
@@ -3,9 +3,10 @@ import { LeaveGroupUseCase } from "../ports/in/use-cases/leave-group.use-case";
 import { LeaveGroupCommand } from "../ports/in/commands/leave-group.command";
 import { GetAuthenticatedUserUseCase, GetAuthenticatedUserUseCaseProvider } from "src/auth/application/ports/in/use-cases/get-authenticated-user.use-case";
 import { GroupRepository, GroupRepositoryProvider } from "../ports/out/group.repository";
-import { UnauthorizedAccessError } from "src/auth/application/services/errors/unauthorized-acess.error";
+import { UnauthorizedAccessError } from "src/auth/application/services/errors/unauthorized-access.error";
 import { GroupMemberRepository, GroupMemberRepositoryProvider } from "../ports/out/group-member.repository";
 import { GroupNotFoundError } from "./errors/group-not-found.error";
+import { UnableToLeaveGroupError } from "./errors/unable-to-leave-group.error";
 
 @Injectable()
 export class LeaveGroupService implements LeaveGroupUseCase 
@@ -30,10 +31,22 @@ export class LeaveGroupService implements LeaveGroupUseCase
             throw new GroupNotFoundError(`Group not found`);
         }
 
-        const groupMember = (await this.groupRepository.findBy({groupId: group.id(), userId: authenticatedUserId})).at(0);
+        const groupMember = (await this.groupMemberRepository.findBy({groupId: group.id(), userId: authenticatedUserId})).at(0);
 
         if(groupMember === null || groupMember === undefined) {
             throw new UnauthorizedAccessError(`You are not a member of the group`);
+        }
+
+        const groupMembers = await this.groupMemberRepository.findBy({groupId: group.id()});
+
+        let hasAdministratorGroupMember = false;
+
+        groupMembers.forEach((gm) => {
+            if(gm.role().name() === 'administrator') hasAdministratorGroupMember = true;
+        });
+
+        if(groupMember.isCreator() || !hasAdministratorGroupMember) {
+            throw new UnableToLeaveGroupError(`Unable to leave group`);
         }
 
         this.groupMemberRepository.delete(groupMember.id());
